pyproject.toml
[tool.poetry]
name = "webctfchallenge"
version = "0.1.0"
description = ""
authors = ["Lucy Hochkamp <git@xyno.systems>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
webctfchallenge = 'webctfchallenge:main'